[Reason] Semicolon is not required after final item in sequence.
Summary:Per request of @jberdine, we no longer require the final
semicolon. There's a small tradeoff that had to be made (no punned,
single item records) - but that case is very rare, and would surely be
caught by the type system.

I also took the opportunity to start a clean directory of type checked
formatting test cases. Many bugs in parsing/printing would be caught by
simple type checking of the AST.

Thanks for the suggesion, Josh. I like this much better.

The next step is to make it so when pretty printing, the final semicolon
is ommitted.

Similarly, it is easy to make it such that the final item in a
module/signature does not require a semicolon.

Test Plan:Added type-checked formatting tests.

Reviewers:@jberdine, @cristianoc

CC: