[Reason] Add ocp-indent test cases.
Summary:`ocp-indent` has a ton of useful edge case tests that would be
good for making sure we've covered the entirely of the grammar. Some of
them were not valid parses, so I had to go through and fix any part that
was not a valid input parse.

Test Plan:

Reviewers:

CC: