[Reason][Emacs]Replace pure ocaml package's syntax with Reason (fork of Rust)
Summary: This takes the emacs editing layer (initially copy of ocaml
layer) and replaces the syntax with a fork of Rust.

Even in this very bare bones state, I've confirmed that:

1. Autocomplete actually works! (Although the types are printed as ML
types not reason types).

2. Jump to location already works! (as long as you have checked out my
merlin pull request, built and `OPAM pin`ned it).

The next steps are to:

1. Rename the emacs layer to be "reason" instead of "ocaml" so it does
not conflict.

2. Make the syntax not Rust! (Indentation of switch statements as you
type and press enter are one issue to solve, and there's a bunch of
easier stuff too).

3. Add a format hook (maybe copy how spacemacs' rust layer does this).o

4. Make merlin not fail on Reason syntax (Frederic has a branch of
merlin that allows us to inject our custom "reader").

Test Plan:

Reviewers:

CC: