[Reason] Properly parse single underscore pattern.
Summary:
We need to properly parse the underscore as being *potentiall* a pattern
for *all* arguments of a constructor. That means in the case of a single
underscore, we need to stop putting the "explicit_arity" attribute on
patterns (normally that "explicit_arity" attribute is awesome, but in
this case we just need to relax things a bit).

Test Plan:See test cases.

Reviewers:yunxing, cristiano

CC: