Introducing an ambiguous syntax for Constructor args
Assuming you have a ML program that looks like this:

```
match a with
  | C (c, d) -> 1
  ;;
```

What information about arguments do you get from constructor C? Is it a constructor that takes a tuple? or is it a constructor that takes two arguments? The compiler has to use heuristics to figure it out but we don't have that information at the parsing time.

This diff introduces this solution:

When translate from ML to Reason, we can provide an ambiguous syntax thats support both cases. The reason ambiguous syntax will be compilable. The user later on has to decide which case he intends to use and gradually migrate from the ambiguous syntax to a specific Reason syntax.

As an example, when an user first convert the above code from ML to Reason. It will be something like this:

```
switch a {
  | C (TODO_REMOVE_AMBIGUITY__ f x __TODO_REMOVE_AMBIGUITY) -> 1
}
```

The above syntax will be valid and he can start compiling it, the pretty printer is idempotent for this code segment.

Later on he can decide which interpretation to convert to, it is either:

```
switch a {
   | C (f, x) -> 1
}
```

Or

```
switch a {
    | C f x -> 1
    }
}
```