Add ocaml-better-errors error reporter to reason
Uses the better error output from
https://github.com/chenglou/ocaml-better-errors

I've made a few modifications to that repo so that it can accept custom
parsers, which, in this case, is a type incompatible error parser that
delegates to BetterError's type incompatible parser, but does some
post-processing to make the type output use the reason syntax rather
than the OCaml one.

This exposes a new terminal command called `rerr` (name temporary). Try
it with `ocamlc someFileWithCompilerError.ml |& rerr`. See that repo for
more details.

**BetterErrors currently isn't submitted to OPAM yet, so please do `opam
pin add BetterErrors` first to make sure this works.**

The names casing in this file are inconsistent, partially because I'm
also confused as to whether we're using snake_case, camelCase or
kebab-case.