[Reason] Recover from lexing errors.
Summary:Before this diff, if you had an error in the lexer, it would
potentially crash the merlin process. This uses the same mechanism we
use for parsing error recovery but for lexing. It actually is inaccurate
for some kinds of lexing errors, but at least it doesn't crash for now.

Test Plan:

Reviewers:@yunxing

CC: