[NuclideReason] Pretty-print type hints as Reason ones
We take the Merlin output and print them through `refmttype`, which
takes a list of types to format.

This refactors `refmttype` (and plugs a few potential bugs) to have an
error fallback step of treating the input not as a core_type AST node
(e.g. `int list`), but as a valid program (e.g. `type a = int list`).
This is because when we hover over a type declaration, Merlin returns
that same type declaration (maybe sometimes in a different way, e.g.
showing the equivalent types? So it's not just an identity operation).
We needed to be able to parse that instead of erroring.

There's a cache for perf.

Test plan:

(Don't forget to `opam upgrade reason` to get the latest `refmttype` CLI
tool.)
Hover over a normal value. Hover over a type declaration. See they all
print correctly into the Reason syntax. Hover over a .rei file. Still
works. Hover over everything.