[Type formatter] Parse module signature
Merlin returns e.g. `sig val x:int end` when, say, we query the type of
`List`. We need to introduce a new entry point to compiler-libs's parser
which starts with module_type. But since that means we need to copy
compiler-libs over and patch it, I'm bailing out by just appending
"module X: " in front of a type.

Also refactored the code to be more streamlined.