[NuclideReason] New method for requiring reason grammar for highlight
For type hint highlighting, we've previously asked people to `npm link`
in language-reason and `npm link language-reason` to install this
language-reason dependency in NuclideReason/node_modules. This is so
that we can require the path globally for atom's Highlight's API (see
the diff comments for more detail). Unfortunately our bundling method
doesn't work with this too much. So here's the alternative. See comments
for more.