[NuclideReason] Make hyperclick (somewhat) work again
There are a few bugs that I suspect originated from the interaction of Merlin and Reason.

For example, clicking a symbol that's defined in the same file doesn't work.

I'll investigate and ask the right people later. But for now, this is strictly better than not working 100% of the time.