[AtomReason] Send protocol version on each command
Fixes #281.
The nuclide-ocaml plugin need to still do its job of sending the
procotol on its side though. But since apparently nobody uses that in
conjunction with the reason plugin we can... maybe defer it. Or I should
unlazy myself and submit a PR to nuclide.

Also, merlin currently has a bug for sending the protocol. For the
pinned hash we're on, it always return `{selected: 2}` for whatever
protocol version we send (which is convenient for us, because we're
actually using version 2). For master, it returns 3 when we ask for 3, 2
for 2, and finally returns 2 when we ask for 1 (bug). Stable opam merlin
  simply fails (@bsansouci).

@jordwalke