[AtomReason] Remove internal dependency on Nuclide
The only place left that's using Nuclide was in the locate (hyperclick)
code. We've inlined the source and removed the dep on Nuclide.

Note that as of right now, `locate` only works for jumping to stdlib
modules. This morning it worked with jumping to another file in our own
project (e.g. in `index.re`, any mention of `Atom.Promose`. Click on it
and it'd jump to Atom.re to the `Promise` module. Though if the mention
is e.g. `Atom.Promise.createFakePromise`, clicking on
`createFakePromise` errors), but it stopped working this afternoon and
for the life of me I can't make it work anymore.