[AtomReason] Implement smart selection of occurrences (#355)
Expose a shortcut and functionality that allows us to select all the
occurrences of the name under cursor. Upon activating the shortcut
(alt-d, akin to cmd-d), every occurrence of the name will be selected.
This is more than just a naive text selection; same variable name in
unrelated scopes won't be selected.