[refmt.el] Use write extention when formatting files
In refmt.el, we copy current buffer into a tmp file and apply refmt to it.
Previously we only copy the buffer as ".re" files, this reports syntax errors
on ".rei" files.

This patch checks the extention of the current buffer, and creates
a tmp file with the right extention.