Fix formatTest under Linux due to incorrect mktemp usage
`mktemp` under Debian has a helpfully different syntax from OSX.
This caused the test to fail with an error message, and the lack
of `set -e` caused it to continue on and attempt to create a bunch
of directories under the root directory.

This activates quit-on-error mode in bash and also a more generic
mktemp syntax that should work on both platforms.

Signed-off-by: Anil Madhavapeddy <anil@recoil.org>