[Reason][Request For Comments] Better indentation for switch cases.
Summary:This diff relaxes the requirement that the right hand side of an
`=>` in `switch` cases must be parsable as an expression in any other
context. Instead, we say that the right hand side of `=>` in a `switch`
case must either be parsable as an expression, or parsable as the
contents of a "sequence(block) expression". This is basically saying that
there are "implicit" sequence braces wrapping the right hand side of the
`=>` so you don't need to add them there.

This saves considerable indentation (two spaces for every level of
`switch` nesting).

Test Plan:

Reviewers:

CC: