[Reason] Make Reason parse locations on Variant types match OCaml's.
OCaml includes the bar in the location of variants and Reason did not.

We'll make Reason match OCaml here because it's always better to include
the bar in the location. There's other places we still need to fix in
Reason to also include the bar in the location (switch cases, fun
cases), which will make it diverge from OCaml but in a good way.