[Reason] Make requirements for infering EOL comments more strict.
Summary:
In order to be considered an end of line comment, we must know that the
beginning of line is exactly at the attachment point of the comment
(recall that the attachpoint is at X in the following exmample)

   X y z       /* end of line comment */

While this is more correct it causes formatting issues because in
reality, we've set the attach point to the | in the following example
(correctly so). The problem is that the OCaml AST doesn't correctly consider
the bar to be part of the match row location. Neither does Reason. So
some of the formatting cases for Reason became worse.  We will fix the
Reason location of match rows in the next diffs, but that won't fix the
OCaml code you're parsing and converting from. For that, we will have to
apply some final hacks such as annotating each comment with whether or
not it was attached to a bar - and if so allows 2 characters of slop.

  | X y z =>..  /* end of line comment */

Test Plan:

Reviewers:

CC: