[Reason] Make Reason Idempotent Again
Summary:A previous diff had removed the idempotent nature of the pretty
printer, by taking

    let x =
      something; /* comment */

and formatting it like this

    let x =
      something /* comment */;

This is fine, if the outer semicolon having term does not begin on the
line, then "something" is the best candidate for having the end of
line comment attached to it. If that's the case, then it will jump
inside of the semicolon to indicate the association. (I think it's
actually a good thing to indicate what was inferred as the end of line
comment, and format accordingly). If we did not indicate this in the
formatting you'd be in for a surprise when joining several lines onto
one, and seeing how the meaning of the end of line comment has changed.
When we have the formatter give feedback about what the eol comment is
attached to, those surprises are eliminated.

The problem, is that before this diff, we were not clever enough to
continue to interpret the above example as an eol comment on
"something", because the semicolon would get in the way - making the
comment appear to not be an eol comment. This diff special cases
terminators such as ; and , so that they are ignored when attributing
the eol comment - we ignore trailing terminators. That necessitated a
large refactoring.

Test Plan:

Reviewers:

CC: