[Reason] Support last row in record type definitions.
Summary: No place in the AST records the locations of the open/close
braces, but we can come up with a heuristic that captures the vast
majority of use cases.

If you write code like the following, (either in OCaml, or Reason) your
comments will be preserved in their correct locations.

```
/* Before type t */
type t = {
  /* Before x */
  x: int,  /* after x */
  /* Before y */
  y: int  /* after y */
  /* Final row */
};
```

Watch out for the following common mistake:

```
type t = {
  /* Before x */
  x: int,  /* after x */
  /* Before y */
  y: int  /* after y */ };

The closing brace position makes `y` not be at the *end* of the line, so
it won't be preserved as an end of line comment.

```

Test Plan:

Reviewers:

CC: