Fixes the following issues with outcometree printing
```
Reason # type t 'a 'b = option ('a, 'b);
type t 'a 'b = option ('a, 'b)
```

used to be printed as `type t 'a 'b = option (('a, 'b))` with
superfluous brackets.

```
Reason # type t 'a 'b = option ('a => 'b);
type t 'a 'b = option ('a => 'b)
```

used to be printed as `option 'a => 'b`, which is incorrect.