Temporary stop-gap until reason can be installed as a findlib package in npm.
Currently, we can only convert `opam` packages to `npm` when they support an `opam-less` installation method. Right now, the Reason install assumes you have `opam-installer`, which we don't have on `npm` yet. So for now, this is a hacky workaround that will unblock people (`rebuild` was broken because our build scripts try to find Reason in `findlib` root, which we cannot install `reason` into without having `opam-installer` (or equivalent)).