Fix #736. Add support for `#` followed by any of `['!' '$' '%' '&' '+' '-' '.' ':' '<' '=' '>' '?' '@' '^' '|' '~']` or `('\\' '*') | ('\\' '/')` to be printed properly with the right precedence and associativity.
Special cased all extension operators to keep the proper associativity.
Example `a##(b##c)` keeps the parens, `(a##b)##c` doesn't need to.

Support for simple_get_application which means we don't need to add
parens when calling functions. For example `add (x##y) (a##b) ` gets
pretty printed as `add x##y a##b`.

This means that `#=` needs parens in the case of the bug associated with
this PR: `ret #= (Some 10)` but the pretty printer will keep them.

Kudos to @schmavery for coming up with this quick solution.