[JSX] JSX: "Simplified" version of Sander's diff.
Summary: This is a version of Sander's JSX implementation which treats
JSX as "simple" constructs, meaning they are "clearly one unit", and do
not require parenthesis to be wrapped around them in order to be passed
as function arguments. That means you can treat <tag ...> as the
"opening paren" and "</tag>" as the closing paren.

This diff also supports:

   - Frags (which are simply lists that don't require commas separating
     contents).

   - Lists of JSX items

```
let myFunction =
   callSomeFunction
     <FirstItem>foo</FirstItem>
     <SecondItem>foo</SecondItem>;

let lst = [<Tag />, <AnotherTag />];

let lst =
  <>
    <Tag />
    <AnotherTag />
  </>;

```

Whats Not Implemented: Currently, the limitation of this approach is
that subsequent JSX elements must always have spaces separating them.
I have a good idea about how to solve that limitation, but we can start
using this right now, in an experimental mode.

```
let items = <Wrapper> <This /> <Must /> <Have /> <Spaces /> </Wrapper>
```

Test Plan:Test cases.

Reviewers:

CC: @vjeux