[Reason] Remove shrinkwrap.
Summary: This actually doesn't work correctly. It was merely never being
tested before (because shrinkwrap files are ignored when a dependency
specifies a shrinkwrap). I tested by doing a git clone of reason and
`npm install` inside the reason directory and it failed.

Yunxing removed the problematic packages, but we'll need a better, more
comprehensive solution.

Test Plan:

Reviewers:

CC: