Clarify Reason npm package is lowercase (#947)
In the 'Install Via `npm` block Reason is referred to in a code block. Since these are code blocks I and I think many others honor the capitalization in them. Reading on it becomes clear the Reason project refers to Reason in a code block always, having nothing to do with the names of different Reason related things such as the npm package. Which is fine, even good I'd say. Seeing `Reason` everywhere and no mention of the package name had me guess it to be 'Reason' which is wrong and unfortunately results in a cryptic npm error: `npm ERR! Cannot read property 'path' of null`.

We can avoid letting the user guess and avoid possibly steering them in the wrong direction (by writing `Reason` in the block) by mentioning the package name. The project does the same in other places in the README where a name for something Reason related is required; I suggest we do the same for the npm block :].

I will also see if I can help improve this experience on the npm side ^_^ .