Remove rebuild hack (#1173)
We had a lot of hacks into rebuild so that we could statically link our
custom ocamlbuild rules into rebuild. It was a pain to maintain and debug.

Max and I found a less hacky approach to link the custom rules defined
in reasonbuild.ml:

1. Move reasonbuild.ml into a tmp directory, call it myocamlbuild.ml
2. Run ocamlbuild with '-just-plugin' option in that directory
3. Step 2 will generate a binary called "myocamlbuild", rename it and ship it
as "rebuild"

After step 3, 'rebuild' can now be used as an augmented 'ocamlbuild' with custom rules.