New jsx ppx tweaks (#1288)
* New jsx ppx tweaks

Accepts a command line flag. If it's `-version 2`, then use the new JSX
transform. If it's anything else, then:
  - use the old JSX transform
  - if the file has `[@@@bs.config {jsx: 2}]` at the top, use the new
    transform.

This makes sure nothing break when we upgrade everyone to the new
transform (used by rr2), *and* that when they start a new project, they
can put a `{"reason": {"react-jsx": 2}}` in `bsconfig.json` and have the
new transform without needing to put any config at the top of the files
(bsb will take care of invoking the ppx with the right flags depending
on bsconfig here).

* New tweaks

* Document new jsx ppx