Ignore some typecheck tests before OCaml version
Introduces a convention that typeCheck tests named like
"testName.4.03.0.re" will ignore the compilation step with ocamlc
on compiler versions lower than the version in the test name.
A test named as in the example would only be compiled on
OCaml 4.03.0 and later.

Moves the let-exception tests into their own file that runs only on
OCaml versions >= 4.04.

Displays a notice in the output if the file's compilation step was
skipped.