Implement "intelligent" breaking for records & objects.
The idea is that a line break in input should signify an intent to format a record/object with line breaks.  Refmt "relaxes" its hard coded rules in favour of checking the input for possible breaks. Since the start & begin location are part of the ast, we simply have to check if the line height of start & end loc are the same. If they are, then everything was on one line.

This has the benefit that `type point = {x: int, y: int}`, can be
printed one line.

type t = {
    a: int, b: int};

type t = {a: int,
  b: int};

type t = {a: int, b: int
};

=> Since the structures above have been put on multiple lines,
everything will be formatted as:

type t = {
  a: int,
  b: int
};