Add support for foo(~x=?).
It seems that the existing parsing of `foo(~x=?)` is not intentional, so it’s been removed.

Syntactically, the existing `foo(~x?)` and `foo(~x=?None)`, are quite similar to `foo(~x=?)` though the latter has a completely different meaning, nothing to do with optional labels.

Alternatively, the underscore “_” symbol is still free in the context where “?” is used here.