WIP initial prototype dynamic precedence of fast pipe.
In general . and -> belong to the same category of
precedence. However in the context of `myAction->self.send`, the
`.` has higher precedence.
This contrasts with `event->target##value` or
`location##streets.foo[1]` where we read the expression from left to right.

Hence we need a way to simulate dynamic precedence.