Attempt to fix the braces removal in JSX props
This is a very empirical attempt, I'm not exactly sure if it's a safe patch.

I basically just modified the printer bit by bit until I found something that'd fix my test case.

I also tried to change the `dont_preserve_braces` in the `Pexp_apply` case in `formatJSXComponent` but it seemed to be a bit more aggressive. The issue does seem to flow down from there though.

Fixes #2471