Fix: Failure("input_value: not a binary channel")
To the ppx @jfrolich output reason to binary and convert back. This was failling on windows. On Unixes open_in and open_in_bin behave the same not on windows accondingly to [this thread](https://www.developpez.net/forums/d861244/autres-langages/langages-fonctionnels/caml/input_value-not-binary-channel/) in french (please use google translate)