Print numeric ranges to a valid parse. Custom error message on invalid parse.
Summary: This prints numeric ranges to a valid parse (even though they
wouldn't type check anyways, it's nice for ppx extension patterns).
Also, I added a custom error message in IDEs for when you forget to include a
space between the numers, but it recovers to something sensible.

Test Plan:

Reviewers:

CC: